Use a derive trait for AsEntity
