+++ Cargo.toml
@@ -1,50 +1,5 @@
-[package]
-edition = "2018"
-name = "bigbang"
-version = "0.0.9"
-authors = ["Alex <alex@alex-hansen.com>"]
-license-file = "LICENSE.md"
-description = "An optimized framework for n-(hard)-body gravitational simulation"
-repository = "https://github.com/sezna/bigbang"
-keywords = ["nbody", "gravity", "simulation"]
-categories = ["simulation"]
-exclude = ["test_files/*"]
-readme = "README.md"
-documentation = "https://docs.rs/bigbang"
-
-[dependencies]
-either = "1.5"
-rayon = "1.3"
-serde = { version = "1.0", features = ["derive"] }
-
-[dev-dependencies]
-criterion = "0.3"
-iron = "0.6"
-router = "0.6"
-staticfile = "0.5.0"
-serde_derive = "1.0.100"
-serde = "1.0.100"
-mount = "0.4.0"
-serde_json = "1.0.40"
-lazy_static = "1.4.0"
-iron-cors = "0.8.0"
-websocket = "0.23.0"
-multiqueue = "0.3.2"
-tokio = "0.1.22"
-tokio-reactor = "0.1.9"
-futures-preview = { version = "=0.3.0-alpha.18", features = ["compat"] }
-futures-util = "0.2.1"
-log = "0.4.8"
-fern = "0.5.8"
-bincode = "1.1.4"
-chrono = "0.4.9"
-futures-timer = "0.3"
-rand = "0.7.0"
-
-[lib]
-crate-type = ["cdylib", "lib", "staticlib"]
-
-[[bench]]
-name = "theta"
-harness = false
-
+[workspace]
+members = [
+	"bigbang",
+	"bigbang_derive"
+]
