+++ bigbang/examples/c_example.c
