+++ bigbang/examples/websocket_frontend/index.js
