+++ bigbang/src/as_entity.rs
