+++ bigbang/src/collisions.rs
