+++ bigbang_derive/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "bigbang_derive"
+version = "0.1.0"
+authors = ["Alex Hansen <alex@alex-hansen.com>"]
+edition = "2018"
+
+[dependencies]
+quote = "1.0"
+syn = { version = "1.0", features = ["derive"] } 
+proc-macro2 = "1.0"
+
+[lib]
+proc-macro = true
